+++ lib/lightning_web/endpoint.ex
@@ -39,6 +39,7 @@ defmodule LightningWeb.Endpoint do
     cookie_key: "request_logger"
 
   plug Plug.RequestId
+  plug PromEx.Plug, prom_ex_module: Lightning.PromEx
   plug Plug.Telemetry, event_prefix: [:phoenix, :endpoint]
 
   plug Plug.Parsers,