+++ test/lightning/prom_ex_test.exs
@@ -1,0 +1,38 @@
+defmodule Lightning.PromExTest do
+  use ExUnit.Case, async: true
+
+  test "returns dashboard config" do
+    Application.put_env(:lightning, Lightning.PromEx, %{datasource_id: "foo"})
+
+    expected = [datasource_id: "foo", default_selected_interval: "30s"]
+
+    assert Lightning.PromEx.dashboard_assigns() == expected
+  end
+
+  test "returns enabled dashboards" do
+    expected = [
+      {:prom_ex, "application.json"},
+      {:prom_ex, "beam.json"},
+      {:prom_ex, "phoenix.json"},
+      {:prom_ex, "ecto.json"},
+      {:prom_ex, "oban.json"},
+      {:prom_ex, "phoenix_live_view.json"}
+    ]
+
+    assert Lightning.PromEx.dashboards() == expected
+  end
+
+  test "returns enabled plugins" do
+    expected = [
+      PromEx.Plugins.Application,
+      PromEx.Plugins.Beam,
+      {PromEx.Plugins.Phoenix,
+       router: LightningWeb.Router, endpoint: LightningWeb.Endpoint},
+      PromEx.Plugins.Ecto,
+      PromEx.Plugins.Oban,
+      PromEx.Plugins.PhoenixLiveView
+    ]
+
+    assert Lightning.PromEx.plugins() == expected
+  end
+end