+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to
   [#1195] (https://github.com/OpenFn/Lightning/issues/1195)
 - Add a OpenTelemetry trace example
   [#1189] (https://github.com/OpenFn/Lightning/issues/1189)
+- Add plumbing to support the use of PromEx
+  [1199](https://github.com/OpenFn/Lightning/issues/1199)
 
 ### Changed